pyproject.toml
[project]
name = "reside"
version = "0.0.1"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "coverage==7.6.1",
    "django==5.1",
    "django-phonenumber-field==8.0.0",
    "jsonfield==3.1.0",
    "phonenumberslite==8.13.42",
    "python-dateutil==2.9.0.post0",
    "six==1.16.0",
    "sqlparse==0.5.1",
    "toposort==1.10",
    "gunicorn==22.0.0",
    "whitenoise==6.7.0",
    "psycopg2-binary==2.9.9",
    "dj-database-url==2.2.0",
    "python-decouple==3.8",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["reside", "users", "tenants", "properties", "payments", "core", "accounts"]

[tool.uv]
package = false

[project.optional-dependencies]
dev = [
    "django-debug-toolbar==4.4.6",
    "django-seed==0.3.1",
    "factory-boy==3.3.0",
    "faker==26.3.0",
    "ruff>=0.6.1",
]
